add import order checks to flake8
Note that local packages need to be specified to flake8 via run-tests.py

This commit makes a small alteration to `.travis.yml` which is also in PR #20.
I added it here as well to get travis to pass for this PR.

I had to make a couple of other small edits as well for flake8. I also updated
a test so RepoPublishConduit was not used.