Remove the doc and coverage checks from run-tests.py.
doc_check and coverage checking are new in Pulp's master branch, but we want 1.0.0 of the
plugins to work against Pulp 2.5 ideally, or at least 2.6.0. This commit removes the doc_check
and coverage check from run-tests.

Also, 2.5 did not list its dependencies in its setup.py, so this adds them to the .travis.yml
file.

This change must not be propagated to master.