Create new model for python packages
This commit does 4 things:
    1. Rename `_filename` to `filename`.
    2. Switch unit key of Package to filename.
    3. Remove unnecessary fields from the model.
    4. Add the machinery to create instances of new model from the PyPI
       JSON metadata and directly from the packages.

The filename contains the name, version, platform, and python versions.
This means that filename is guaranteed to be unique on PyPI. This is
necessary because `name` and `version` will no longer guarantee
uniqueness when we are supporting multiple package types. For instance,
two seperate packages, a wheel and and sdist can and would share the
same name and version.

There are two reasons why I chose to drop the extra fields. Firstly, the
expected use case will be that users will specify which packages they
want to install from a pulp python repository, they will not need to
search the repo to figure out which packages to install. This means that
the fields on the model that are not necessary for sync or publish are
not very useful. The second problem is that if we are to implement lazy
sync for python, we need to be able to create the units from the
metadata available on PyPI. This metadata combines some fields like
`license` and `home_page` for all versions, platforms, and package types
of a given package name. This could present an issue where some of this
information changed over time, but the newest version of that data is
applied to all units. Since it is not really necessary in the first
place, it is preferred to leave the information out than to possibly
present inaccurate information.

The field `summary`, though not strictly necessary, provide helpful
information and was left for convinence, despite the possibility that
it will change over time.

closes #1882

mor minimal model

mor docs