Publish metadata to allow pulp to pulp sync
Publishes a json metadata file for each python project that contains
metadata for each python package that belongs to it. This is a subset of
the api that PyPI implements, and allows a Pulp to sync from another
Pulp in exactly the same way that it would sync from PyPI itself.

closes #140