Remove intersphinx and enable Strict mode
Intersphinx was preventing the strict mode from being enabled
due to docs being built in a network isolated mock environment.

Intersphinx was also going to break because it is tied
to URLs from ReadTheDocs. Now intersphinx is removed, and
usage of it was dropped from the content. Strict Sphinx
mode is now enabled.

https://pulp.plan.io/issues/2034
re #2034