Serve metadata files at directory
PyPI's API acts like a live API. Requests are made to urls that end in
'/'. Since publishing actually creates these files, we served them at
'/index.html' in Pulp 2, and made the adjustment in the Apache
configuration. Since we would like to be webserver agnostic, that isn't
a good solution for Pulp 3.

Instead, we simply set the path to serve to be different from the path
of the file.

https://pulp.plan.io/issues/3465
fixes #3465