Only use qualified noqa pragmas, and as needed
Don't use the bare `noqa` pragma. Doing so instructs flake8 to ignore
errors other than the one being targeted. In addition, only use the
`noqa` pragma on lines of code that actually trigger a complaint. As a
case in point, consider this code:

    from X import (gen_remote, gen_publisher,   # noqa
                   set_up_module as setUpModule)

This pragma is intended to ignore the fact that `setUpModule` is not
used. However, it ends up ignoring all possible errors triggered by the
two lines of code, including the fact that `gen_publisher` isn't used.
(In some modules.)

This kind of change is especially important given the great number of
changes that Pulp Smash is likely to undergo in the near future.