Update distribution creation to use task_handler
Distribution creation was modified to be an async task.
A new response_handler was added to pulp-smash to deal with those cases.
Update distribution creation to use this new response_handler.

Also fixed failing tests surrounding change in the default for mirror from
True to False when syncing remotes.

'[noissue]'