Expand and connect the docs
With the addition of published REST API reference docs, we have a lot
more room for well-connected thorough documentation. These docs are
generated from comments in the Viewsets, and are rendered in html, which
gives us more options for format and clarity.

In this PR I've made sure that each `workflow` corresponds to a major
feature, minor features are all clearly explained in the autogenerated
docs, and that they are connected for a seamless experience.

[noissue]