Fix REST API link on top level index
Locally, the link worked prior to this change, but the link is dead on
rtd. I suspect this has something to do with the namespacing there, ie
"latest". We removed rtd theme from the local build config recently to
address a different linking issue.

NOTE: this commit is not really testable until it is merged.

[noissue]